[REM-103] refactor: clean up and reorganize MasterApp and MenuManager
- crawjud/__init__.py: removed unused worker and server methods, updated menu options
- crawjud/menu_manager.py: added properties for Celery and Quart instances, reorganized methods for clarity
