+++ crawjud/__init__.py
@@ -2,25 +2,19 @@
 
 import asyncio
 from pathlib import Path
-from platform import node
-from threading import Event, Thread
 from time import sleep
 
 import inquirer
 from billiard.context import Process
 from celery.apps.beat import Beat
-from celery.apps.worker import Worker
 from clear import clear
-from socketio import AsyncServer  # noqa: F401
+from socketio import AsyncServer
 from termcolor import colored
 from tqdm import tqdm
-from uvicorn import Server  # noqa: F401
 
-from crawjud.config import StoreThread, running_servers
+from crawjud.config import running_servers
 from crawjud.core import create_app
-from crawjud.core.watch import monitor_log
 from crawjud.menu_manager import MenuManager
-from crawjud.utils.gen_seed import worker_name_generator
 
 io = AsyncServer(
     async_mode="asgi",
@@ -58,71 +52,6 @@ class MasterApp(MenuManager):
         Process(target=start_beat, daemon=True).start()
         clear()
 
-    def start_worker(self, stop_event: Event) -> None:
-        """Run the Celery worker in a thread controlled by a stop event.
-
-        Args:
-            stop_event (Event): Event to signal the thread to stop.
-
-        """
-        worker_name = f"{worker_name_generator}@{node()}"
-        worker = Worker(
-            app=self.celery,
-            hostname=worker_name,
-            task_events=True,
-            loglevel="INFO",
-            concurrency=50.0,
-            pool="threads",
-        )
-        Thread(target=worker.start, name="Worker Celery").start()
-        while not stop_event.is_set():
-            sleep(1)
-
-        worker.stop()
-
-    def start_quart(self, stop_event: Event) -> None:
-        """Run the Quart server in a thread controlled by a stop event.
-
-        Args:
-            stop_event (Event): Event to signal the thread to stop.
-
-        """
-        while not stop_event.is_set():
-            ...
-
-    def start_all(self) -> None:
-        """Start all server components in separate threads and allow stopping with an event.
-
-        This method creates threads for the worker, Quart server, and Celery beat.
-        It listens for a keyboard interrupt and then signals all threads to stop.
-        """
-        stop_event = Event()
-        worker_thread = Thread(target=self.start_worker, args=(stop_event,))
-        quart_thread = Thread(target=self.start_quart, args=(stop_event,))
-
-        worker_thread.start()
-        quart_thread.start()
-
-        store_thread = StoreThread(
-            process_name="Worker",
-            process_id=worker_thread.ident,
-            process_status="Running",
-        )
-
-        running_servers["Worker"] = store_thread
-
-    def status(self) -> None:
-        """Log the status of the server."""
-        if not running_servers.get("Quart API"):
-            return ["Server not running.", "ERROR", "red"]
-
-        clear()
-        tqdm.write("Type 'ESC' to exit.")
-
-        monitor_log("uvicorn_api.log")
-
-        return ["Exiting logs.", "INFO", "yellow"]
-
     def prompt(self) -> None:
         """Prompt the user for server options."""
         self.current_menu_name = "Main Menu"
@@ -147,7 +76,7 @@ class MasterApp(MenuManager):
                 clear()
 
             menu = {
-                "Quart API ASGI": self.quart_menu_api,
+                "Quart Application": self.quart_application,
                 "Celery Worker": self.worker_menu,
                 "Celery Beat": self.beat_menu,
             }
